+++ src/app/(beforeLogin)/_lib/signup.ts
@@ -42,7 +42,7 @@ export default async (
   }
 
   if (shouldRedirect) {
-    redirect("/home"); // try/catch문 안에서 X
+    redirect("/home");
   }
   return { message: null };
 };
